updaet typography components & docs
