+++ packages/web-ui/src/Heading/Heading.docs.mdx
@@ -8,6 +8,9 @@ import * as Stories from './Heading.stories';
 
 Heading renders the primary UW font, to be used for heading-level typography.
 
+The system props are not available on the Heading component, if you need custom typography then use the [Typography](?path=/docs/components-typography--documentation) component.
+If necessary you can use `sx` as an escape hatch for one-off custom styling.
+
 ## Alternatives
 
 - [Text](?path=/docs/components-text--documentation) - For body text
