+++ packages/web-ui/src/Heading/index.ts
@@ -1,4 +1,5 @@
 export { default, headingVariantMapping } from './Heading';
+export { headingThemeOverrides } from './Heading.theme';
 export type {
   DefaultHeadingComponent,
   CustomHeadingProps,
