+++ packages/web-ui/src/Text/Text.docs.mdx
@@ -8,6 +8,9 @@ import * as Stories from './Text.stories';
 
 Text renders the secondary UW font, Work Sans, to be used for body copy.
 
+The system props are not available on the Text component, if you need custom typography then use the [Typography](?path=/docs/components-typography--documentation) component.
+If necessary you can use `sx` as an escape hatch for one-off custom styling.
+
 ## Alternatives
 
 - [Heading](?path=/docs/components-heading--documentation) - For heading-level text
