+++ packages/web-ui/src/Text/index.ts
@@ -1,2 +1,3 @@
 export { default, textVariantMapping } from './Text';
+export { textThemeOverrides } from './Text.theme';
 export type { DefaultTextComponent, CustomTextProps, TextTypeMap, TextProps } from './Text';
