+++ packages/web-ui/src/Typography/Typography.docs.mdx
@@ -11,7 +11,12 @@ Most of the time you should be using `Text` or `Heading`. Typography is not
 affected by any Box context and will not change foreground colour acording to
 the containing Box background prop value.
 
-All the [MUI system props](https://mui.com/system/properties/) are available on this component.
+## Props
+
+All the [MUI system props](https://mui.com/system/properties/) are available on
+this component. The `variant` prop is currently available for backward
+compatability with Customer Web UI, but is deprecated and will be removed in
+the next major version, in favour of the Text & Heading components.
 
 ## Alternatives
