+++ packages/web-ui/src/Typography/index.ts
@@ -1,5 +1,4 @@
 export { default } from './Typography';
-export { typographyThemeOverrides } from './Typography.theme';
 export type {
   DefaultTypographyComponent,
   CustomTypographyProps,
